backend/pyproject.toml
[project]
name = "sensitive-information-audit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.123.5",
    "oracledb>=3.4.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
